+++ opennmt/decoders/decoder.py
@@ -159,6 +159,13 @@ class Decoder(object):
     _ = embedding
     if sampling_probability is not None:
       raise ValueError("Scheduled sampling is not supported by this decoder")
+    if (not self.support_multi_source
+        and memory is not None
+        and tf.contrib.framework.nest.is_sequence(memory)):
+      raise ValueError("Multiple source encodings are passed to this decoder "
+                       "but it does not support multi source context. You should "
+                       "instead configure your encoder to merge the different "
+                       "encodings.")
 
     returned_values = self.decode_from_inputs(
         inputs,
@@ -329,11 +336,13 @@ class Decoder(object):
       output_layer = build_output_layer(self.output_size, vocab_size, dtype=dtype)
 
     state = {"decoder": initial_state}
-    if self.support_alignment_history and not isinstance(memory, (tuple, list)):
+    if self.support_alignment_history and not tf.contrib.framework.nest.is_sequence(memory):
       state["attention"] = tf.zeros([batch_size, 0, tf.shape(memory)[1]], dtype=dtype)
 
     def _symbols_to_logits_fn(ids, step, state):
-      inputs = embedding_fn(ids[:, -1])
+      if ids.shape.ndims == 2:
+        ids = ids[:, -1]
+      inputs = embedding_fn(ids)
       returned_values = step_fn(step, inputs, state["decoder"], mode)
       if self.support_alignment_history:
         outputs, state["decoder"], attention = returned_values
@@ -352,7 +361,8 @@ class Decoder(object):
           decode_length=maximum_iterations,
           state=state,
           return_state=True,
-          min_decode_length=minimum_length)
+          min_decode_length=minimum_length,
+          last_step_as_input=True)
     else:
       outputs, log_probs, state = beam_search.beam_search(
           _symbols_to_logits_fn,
@@ -369,6 +379,7 @@ class Decoder(object):
       lengths = tf.not_equal(outputs, 0)
       lengths = tf.cast(lengths, tf.int32)
       lengths = tf.reduce_sum(lengths, axis=-1) - 1  # Ignore </s>
+      outputs = outputs[:, :, 1:]  # Ignore <s>.
 
     attention = state.get("attention")
     if beam_width == 1:
@@ -378,8 +389,6 @@ class Decoder(object):
       log_probs = tf.expand_dims(log_probs, 1)
       if attention is not None:
         attention = tf.expand_dims(attention, 1)
-
-    outputs = outputs[:, :, 1:]  # Ignore <s>.
 
     if return_alignment_history:
       return (outputs, state["decoder"], lengths, log_probs, attention)
@@ -440,7 +449,8 @@ def greedy_decode(symbols_to_logits_fn,
                   decode_length=None,
                   state=None,
                   return_state=False,
-                  min_decode_length=0):
+                  min_decode_length=0,
+                  last_step_as_input=False):
   """Greedily decodes from :obj:`initial_ids`.
 
   Args:
@@ -449,67 +459,76 @@ def greedy_decode(symbols_to_logits_fn,
     initial_ids: Ids to start off the decoding, this will be the first thing
         handed to symbols_to_logits_fn.
     eos_id: ID for end of sentence.
-    decode_length: Maximum number of steps to decode for.
+    decode_length: Maximum number of steps to decode for (EOS included).
     states: A dictionnary of (possibly nested) decoding states.
     return_state: If ``True``, also return the updated decoding states.
     min_decode_length: Minimum length of decoded hypotheses (EOS excluded).
+    last_step_as_input: If ``True``, only feed the last predicted ids into
+      :obj:`symbols_to_logits_fn`.
 
   Returns:
     A tuple with the decoded output, the decoded lengths, the log probabilities,
     and the decoding states (if :obj:`return_state` is ``True``).
   """
-  batch_size = tf.shape(initial_ids)[0]
-  finished = tf.zeros([batch_size], dtype=tf.bool)
-  step = tf.constant(0)
-  inputs = tf.expand_dims(initial_ids, 1)
-  lengths = tf.zeros([batch_size], dtype=tf.int32)
-  log_probs = tf.zeros([batch_size])
 
-  def _condition(unused_step, finished, unused_inputs,
-                 unused_lengths, unused_log_probs, unused_state):
-    return tf.logical_not(tf.reduce_all(finished))
+  def _condition(step, finished, unused_inputs, unused_outputs,
+                 unused_lengths, unused_cum_log_probs, unused_state):
+    cond = tf.logical_not(tf.reduce_all(finished))
+    if decode_length is not None:
+      cond = tf.logical_and(cond, step < decode_length)
+    return cond
 
-  def _body(step, finished, inputs, lengths, log_probs, state):
+  def _body(step, finished, inputs, outputs, lengths, cum_log_probs, state):
+    # Run next step.
     logits, state = symbols_to_logits_fn(inputs, step, state)
-    probs = tf.nn.log_softmax(tf.to_float(logits))
+    log_probs = tf.nn.log_softmax(tf.cast(logits, tf.float32))
     if min_decode_length > 0:
-      probs = tf.cond(
+      log_probs = tf.cond(
           step < min_decode_length,
-          true_fn=lambda: beam_search.penalize_token(probs, end_id),
-          false_fn=lambda: probs)
+          true_fn=lambda: beam_search.penalize_token(log_probs, end_id),
+          false_fn=lambda: log_probs)
 
     # Sample best prediction.
-    sample_ids = tf.argmax(probs, axis=-1, output_type=inputs.dtype)
-    sample_probs = tf.reduce_max(probs, axis=-1)
+    sampled_log_probs, sampled_ids = tf.nn.top_k(log_probs, k=1)
+    sampled_log_probs = tf.squeeze(sampled_log_probs, axis=1)
+    sampled_ids = tf.squeeze(sampled_ids, axis=1)
+    outputs = outputs.write(step, sampled_ids)
 
     # Don't update finished batches.
-    masked_lengths_inc = 1 - tf.cast(finished, lengths.dtype)
-    masked_probs = sample_probs * (1.0 - tf.cast(finished, sample_probs.dtype))
+    lengths += 1 - tf.cast(finished, lengths.dtype)
+    cum_log_probs += sampled_log_probs * (1.0 - tf.cast(finished, sampled_log_probs.dtype))
+    finished = tf.logical_or(finished, tf.equal(sampled_ids, end_id))
+    if last_step_as_input:
+      next_inputs = sampled_ids
+    else:
+      next_inputs = tf.concat([inputs, tf.expand_dims(sampled_ids, 1)], axis=1)
+    return step + 1, finished, next_inputs, outputs, lengths, cum_log_probs, state
 
-    step = step + 1
-    inputs = tf.concat([inputs, tf.expand_dims(sample_ids, 1)], -1)
-    lengths += masked_lengths_inc
-    log_probs += masked_probs
-    finished = tf.logical_or(finished, tf.equal(sample_ids, end_id))
-    if decode_length is not None:
-      finished = tf.logical_or(finished, step >= decode_length)
+  batch_size = tf.shape(initial_ids)[0]
+  step = tf.constant(0)
+  finished = tf.zeros([batch_size], dtype=tf.bool)
+  outputs = tf.TensorArray(initial_ids.dtype, size=0, dynamic_size=True)
+  lengths = tf.zeros([batch_size], dtype=tf.int32)
+  cum_log_probs = tf.zeros([batch_size], dtype=tf.float32)
+  if not last_step_as_input:
+    initial_ids = tf.expand_dims(initial_ids, 1)
 
-    return step, finished, inputs, lengths, log_probs, state
-
-  _, _, outputs, lengths, log_probs, state = tf.while_loop(
+  _, _, _, outputs, lengths, cum_log_probs, state = tf.while_loop(
       _condition,
       _body,
-      loop_vars=(step, finished, inputs, lengths, log_probs, state),
+      loop_vars=(step, finished, initial_ids, outputs, lengths, cum_log_probs, state),
       shape_invariants=(
-          tf.TensorShape([]),
+          step.get_shape(),
           finished.get_shape(),
-          tf.TensorShape([None, None]),
+          tf.TensorShape([None] if last_step_as_input else [None, None]),
+          tf.TensorShape(None),
           lengths.get_shape(),
-          log_probs.get_shape(),
+          cum_log_probs.get_shape(),
           tf.contrib.framework.nest.map_structure(
               beam_search.get_state_shape_invariants, state)),
       parallel_iterations=1)
 
+  outputs = tf.transpose(outputs.stack())
   if return_state:
-    return outputs, lengths, log_probs, state
-  return outputs, lengths, log_probs
+    return outputs, lengths, cum_log_probs, state
+  return outputs, lengths, cum_log_probs