+++ opennmt/utils/vocab.py
@@ -49,7 +49,7 @@ class Vocab(object):
       filename: The file to load from.
       tokenizer: A callable to tokenize a line of text.
     """
-    with open(filename, "rb") as text:
+    with tf.gfile.GFile(filename, mode="rb") as text:
       for line in text:
         line = tf.compat.as_text(line.strip())
         if tokenizer:
@@ -65,7 +65,7 @@ class Vocab(object):
     Args:
       path: The path where the vocabulary will be saved.
     """
-    with open(path, "wb") as vocab:
+    with tf.gfile.GFile(path, mode="wb") as vocab:
       for token in self._id_to_token:
         vocab.write(tf.compat.as_bytes(token))
         vocab.write(b"\n")
@@ -76,10 +76,9 @@ class Vocab(object):
     Args:
       path: The path to the vocabulary to load.
     """
-    with open(path, "rb") as vocab:
+    with tf.gfile.GFile(path, mode="rb") as vocab:
       for token in vocab:
-        token = token.strip()
-        self.add(tf.compat.as_text(token))
+        self.add(tf.compat.as_text(token[:-1]))
 
   def add(self, token):
     """Adds a token or increases its frequency.